% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distances.R
\name{lrt_distance}
\alias{lrt_distance}
\title{Likelihood Ratio Test (LRT) Distance}
\usage{
lrt_distance(mu1, Sigma1, mu2, Sigma2)
}
\arguments{
\item{mu1}{Mean vector of the first distribution}

\item{Sigma1}{Covariance matrix of the first distribution}

\item{mu2}{Mean vector of the second distribution}

\item{Sigma2}{Covariance matrix of the second distribution}
}
\value{
The LRT distance (non-negative scalar)
}
\description{
Computes a distance inspired by the likelihood ratio test statistic for
multivariate normal distributions. For univariate normals, this distance
is equivalent to the Fisher-Rao distance when applied to maximum likelihood
estimates from samples of equal size.
}
\details{
The LRT distance is defined as:
\deqn{D_{LRT} = \frac{\sqrt{2}}{2} \text{acosh}(2 \cdot LRTMV - 1)}
where
\deqn{LRTMV = \frac{|\frac{1}{2}(\Sigma_1 + \Sigma_2) + \frac{1}{2}(\mu_1-\mu_2)(\mu_1-\mu_2)^T|^2}{|\Sigma_1||\Sigma_2|}}

For univariate distributions with sample size n, the relationship to the
Fisher-Rao distance is:
\deqn{D_{Fisher-Rao} = \frac{\sqrt{2}}{2} \text{acosh}(2 \cdot LRT^{-2/n} - 1)}
}
\examples{
mu1 <- c(0, 0)
Sigma1 <- diag(2)
mu2 <- c(1, 1)
Sigma2 <- matrix(c(2, 0.5, 0.5, 1), 2, 2)
lrt_distance(mu1, Sigma1, mu2, Sigma2)

}
\references{
Eriksen, P. S. (1987). Geodesics connected with the Fisher metric on
the multivariate normal manifold. Proceedings of the GST Workshop.
}
