% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_distances.R
\name{plot_lrt_vs_fisher_rao}
\alias{plot_lrt_vs_fisher_rao}
\title{Plot LRT Distance vs Fisher-Rao Distance Along a Geodesic}
\usage{
plot_lrt_vs_fisher_rao(mu1, Sigma1, mu2, Sigma2, n_steps = 50)
}
\arguments{
\item{mu1}{Mean vector of the first distribution}

\item{Sigma1}{Covariance matrix of the first distribution}

\item{mu2}{Mean vector of the second distribution}

\item{Sigma2}{Covariance matrix of the second distribution}

\item{n_steps}{Number of points along the geodesic (default: 50)}
}
\value{
A ggplot2 object showing LRT distance vs Fisher-Rao distance
}
\description{
Generates points along the Fisher-Rao geodesic between two multivariate normal
distributions and plots the LRT distance against the Fisher-Rao distance at each point.
}
\examples{
\dontrun{
mu1 <- rep(0, 3)
Sigma1 <- diag(3)
mu2 <- rep(1, 3)
Sigma2 <- matrix(c(1, .1, .1, .1, 1, .1, .1, .1, 1), 3, 3)
plot_lrt_vs_fisher_rao(mu1, Sigma1, mu2, Sigma2)
}

}
